Move TweepyStub to own file
